- Merge API (6908-7387) and Z (6911-7414)

+++ v2.7.Z.API/app/Plugin/Tools/Controller/TemplateController.php
@@ -3,12 +3,14 @@
 /**
  * Class TemplateController
  */
-class TemplateController extends AppController
+class TemplateController extends ToolsAppController
 {
 
     public $uses = array(
         'Tools.Template',
         'Tools.TemplateNode',
+        'Tools.CollectionProtocol',
+        'Tools.CollectionProtocolVisit',
         'Group'
     );
 
@@ -18,12 +20,37 @@ class TemplateController extends AppController
         $this->AtimAuth->actionPath = 'controllers/';
     }
 
+    /**
+     * List user collection templates and protocols.
+     */
+    public function listProtocolsAndTemplates()
+    {
+        // Nothing to do
+        
+        // CUSTOM CODE: FORMAT DISPLAY DATA
+        $hookLink = $this->hook('format');
+        if ($hookLink) {
+            require ($hookLink);
+        }
+    }
+
+    /**
+     * List user collection templates.
+     */
     public function index()
     {
-        $this->set('atimMenu', $this->Menus->get('/Tools/Template/index'));
+        $this->set('atimMenu', $this->Menus->get('/Tools/Template/listProtocolsAndTemplates'));
         $this->Structures->set('template');
         
-        $this->request->data = $this->Template->getTemplates('template edition');
+        $templates = $this->Template->getTools('list all');
+        $templateIds = array();
+        foreach ($templates as $newTemplate) {
+            $templateIds[] = $newTemplate['Template']['id'];
+        }
+        $conditions = array(
+            'Template.id' => $templateIds
+        );
+        $this->request->data = $this->paginate($this->Template, $conditions);
         
         $hookLink = $this->hook('format');
         if ($hookLink) {
@@ -31,11 +58,12 @@ class TemplateController extends AppController
         }
     }
 
-    /*
-     * Create the Collection Template setting properties
+    /**
+     * Create a new Collection Template
      */
     public function add()
     {
+        $this->set('atimMenu', $this->Menus->get('/Tools/Template/listProtocolsAndTemplates'));
         $this->Structures->set('template');
         
         // CUSTOM CODE: FORMAT DISPLAY DATA
@@ -47,7 +75,19 @@ class TemplateController extends AppController
         if (! empty($this->request->data)) {
             $submittedDataValidates = true;
             
-            $this->Template->setOwnerAndVisibility($this->request->data);
+            $this->Template->setOwner($this->request->data);
+            $this->request->data['Template']['user_id'] = AppController::getInstance()->Session->read('Auth.User.id');
+            $this->request->data['Template']['group_id'] = AppController::getInstance()->Session->read('Auth.User.group_id');
+            $this->Template->addWritableField(array(
+                'user_id',
+                'group_id'
+            ));
+            if ($this->request->data['Template']['flag_active']) {
+                $this->request->data['Template']['last_activation_date'] = date('Y-m-d');
+                $this->Template->addWritableField(array(
+                    'last_activation_date'
+                ));
+            }
             
             $hookLink = $this->hook('presave_process');
             if ($hookLink) {
@@ -61,11 +101,10 @@ class TemplateController extends AppController
         }
     }
 
-    /*
-     * Build The Collection Template
-     */
     /**
-     * @param $templateId
+     * Modify a Collection Template
+     *
+     * @param integer $templateId Id of the template
      */
     public function edit($templateId)
     {
@@ -79,10 +118,11 @@ class TemplateController extends AppController
         // > 0 -> node in database
         
         // validate access
-        $tmpTemplate = $this->Template->getTemplates('template edition', $templateId);
-        if (empty($tmpTemplate)) {
-            $this->atimFlashWarning(__('you do not own that template'), '/Tools/Template/index/');
-            return;
+        $templateData = $this->Template->getOrRedirect($templateId);
+        
+        $tmpTemplate = $this->Template->getTools('edition', $templateId);
+        if (! $tmpTemplate['Template']['allow_properties_edition']) {
+            AppController::addWarningMsg(__('you do not own that template'));
         }
         
         // js menus required data-------
@@ -135,10 +175,11 @@ class TemplateController extends AppController
             // record the tree
             if ($this->request->is('ajax')) {
                 // ajax request are made to save the template info
-                // TODO validate this section is usefull or not
                 Configure::write('debug', 0);
                 $this->request->data['Template']['id'] = $templateId;
                 $this->set('isAjax', true);
+            } elseif (! $tmpTemplate['Template']['allow_properties_edition']) {
+                AppController::addWarningMsg(__('data can not be changed'));
             } else {
                 // non ajax is made to save the tree
                 $tree = json_decode('[' . $this->request->data['tree'] . ']');
@@ -148,20 +189,26 @@ class TemplateController extends AppController
                 
                 $this->TemplateNode->checkWritableFields = false;
                 foreach ($tree as $node) {
+                    if (! empty($node->defaultValues)) {
+                        $node->defaultValues = $this->TemplateNode->fromateNodeDefaultValuesToSave($node->defaultValues);
+                    }
                     if ($node->nodeId <= 0) {
                         // create the node in Db
                         $parentId = null;
-                        if (isset ($node->parentId) && !is_string($node->parentId)){
+                        if (is_numeric($node->parentId) && is_string($node->parentId)) {
+                            $node->parentId = (int) $node->parentId;
+                        }
+                        if (isset($node->parentId) && ! is_string($node->parentId)) {
                             if ($node->parentId <= 0) {
                                 $parentId = $nodesMapping[$node->parentId];
-                            }elseif ($node->parentId > 0) {
+                            } elseif ($node->parentId > 0) {
                                 $parentId = $node->parentId;
                             }
                         }
                         $this->TemplateNode->data = array();
                         $this->TemplateNode->id = null;
                         
-                        $this->TemplateNode->save(array(
+                        $templateNode = array(
                             'TemplateNode' => array(
                                 'template_id' => $templateId,
                                 'parent_id' => $parentId,
@@ -169,17 +216,27 @@ class TemplateController extends AppController
                                 'control_id' => abs($node->controlId),
                                 'quantity' => $node->quantity
                             )
-                        ));
+                        );
+                        
+                        if (! empty($node->defaultValues)) {
+                            $templateNode['TemplateNode']['default_values'] = $node->defaultValues;
+                        }
+                        $this->TemplateNode->save($templateNode);
                         $nodesMapping[$node->nodeId] = $this->TemplateNode->id;
                         $foundNodes[] = $this->TemplateNode->id;
                     } else {
                         $foundNodes[] = $node->nodeId;
                         $this->TemplateNode->id = $node->nodeId;
-                        $this->TemplateNode->save(array(
+                        
+                        $templateNode = array(
                             'TemplateNode' => array(
                                 'quantity' => $node->quantity
                             )
-                        ));
+                        );
+                        if (! empty($node->defaultValues)) {
+                            $templateNode['TemplateNode']['default_values'] = $node->defaultValues;
+                        }
+                        $this->TemplateNode->save($templateNode);
                     }
                 }
                 
@@ -203,21 +260,23 @@ class TemplateController extends AppController
                 return;
             }
         }
-
+        
         // loading tree and setting variables
         $this->Template->id = $templateId;
-        $this->request->data = $tmpTemplate;
+        $this->request->data = $templateData;
         $this->set('editProperties', $tmpTemplate['Template']['allow_properties_edition']);
-        
-        $tree = $this->Template->init();
+        $tree = $this->Template->init($this->Structures);
         $this->set('treeData', $tree['']);
         $this->set('templateId', $templateId);
-        $this->set('atimMenu', $this->Menus->get('/Tools/Template/index'));
+        $this->set('atimMenu', $this->Menus->get('/Tools/Template/listProtocolsAndTemplates'));
         $jsData = array(
             'sample_controls' => $sampleControls,
             'samples_relations' => $samplesRelations,
             'aliquot_controls' => AppController::defineArrayKey($aliquotControls, "AliquotControl", "id", true),
-            'aliquot_relations' => AppController::defineArrayKey($aliquotControls, "AliquotControl", "sample_control_id")
+            'aliquot_relations' => AppController::defineArrayKey($aliquotControls, "AliquotControl", "sample_control_id"),
+            'fomated_nodes_default_values' => $this->TemplateNode->formatTemplateNodeDefaultValuesForDisplay($templateId),
+            'default_values_json' => $this->TemplateNode->getDefaultValues($templateId),
+            'template_id' => $templateId
         );
         $this->set('jsData', $jsData);
         $this->set('templateId', $templateId);
@@ -228,18 +287,22 @@ class TemplateController extends AppController
     }
 
     /**
-     * @param $templateId
+     * Modify properties of a Collection Template
+     *
+     * @param integer $templateId Id of the template
      */
     public function editProperties($templateId)
     {
-        $templateData = $this->Template->getTemplates('template edition', $templateId);
-        if (empty($templateData))
+        $templateData = $this->Template->getTools('edition', $templateId);
+        if (empty($templateData)) {
             $this->redirect('/Pages/err_plugin_no_data?method=' . __METHOD__ . ',line=' . __LINE__, null, true);
+        }
         if (! $templateData['Template']['allow_properties_edition']) {
-            $this->atimFlashWarning(__('you do not own that template'), '/Tools/Template/index/');
+            $this->atimFlashWarning(__('data can not be changed'), '/Tools/Template/edit/' . $templateId);
             return;
         }
         
+        $this->set('atimMenu', $this->Menus->get('/Tools/Template/listProtocolsAndTemplates'));
         $this->set('atimMenuVariables', array(
             'Template.id' => $templateId
         ));
@@ -255,7 +318,14 @@ class TemplateController extends AppController
         } else {
             $submittedDataValidates = true;
             
-            $this->Template->setOwnerAndVisibility($this->request->data, $templateData['Template']['created_by']);
+            $this->Template->setOwner($this->request->data);
+            
+            if (! $templateData['Template']['flag_active'] && $this->request->data['Template']['flag_active']) {
+                $this->request->data['Template']['last_activation_date'] = date('Y-m-d');
+                $this->Template->addWritableField(array(
+                    'last_activation_date'
+                ));
+            }
             
             $hookLink = $this->hook('presave_process');
             if ($hookLink) {
@@ -276,15 +346,20 @@ class TemplateController extends AppController
     }
 
     /**
-     * @param $templateId
+     * Delete a Collection Template
+     *
+     * @param integer $templateId Id of the template
      */
     public function delete($templateId)
     {
-        $templateData = $this->Template->getTemplates('template edition', $templateId);
-        if (empty($templateData))
+        $templateData = $this->Template->getOrRedirect($templateId);
+        
+        $templateData = $this->Template->getTools('edition', $templateId);
+        if (empty($templateData)) {
             $this->redirect('/Pages/err_plugin_no_data?method=' . __METHOD__ . ',line=' . __LINE__, null, true);
+        }
         if (! $templateData['Template']['allow_properties_edition']) {
-            $this->atimFlashWarning(__('you do not own that template'), '/Tools/Template/index/');
+            $this->atimFlashWarning(__('data can not be changed'), '/Tools/Template/edit/' . $templateId);
             return;
         }
         
@@ -297,11 +372,53 @@ class TemplateController extends AppController
             )
         ));
         $nodesToDelete = array_reverse($nodesToDelete);
-        foreach ($nodesToDelete as $nodeToDelete) {
-            $this->TemplateNode->delete($nodeToDelete);
-        }
-        $this->Template->delete($templateId);
         
-        $this->atimFlash(__('your data has been deleted'), '/Tools/Template/index/');
+        // Check deletion is allowed
+        $arrAllowDeletion = $this->Template->allowDeletion($templateId);
+        
+        $hookLink = $this->hook('delete');
+        if ($hookLink) {
+            require ($hookLink);
+        }
+        
+        if ($arrAllowDeletion['allow_deletion']) {
+            
+            foreach ($nodesToDelete as $nodeToDelete) {
+                $this->TemplateNode->delete($nodeToDelete);
+            }
+            $this->Template->delete($templateId);
+            
+            $hookLink = $this->hook('postsave_process');
+            if ($hookLink) {
+                require ($hookLink);
+            }
+            
+            $this->atimFlash(__('your data has been deleted'), '/Tools/Template/listProtocolsAndTemplates/');
+        } else {
+            $this->atimFlashWarning(__($arrAllowDeletion['msg']), '/Tools/Template/edit/' . $templateId);
+        }
+    }
+
+    /**
+     * Form to record template node default values.
+     *
+     * @param integer $nodeId Id of the template node
+     * @param integer $datamartStructureId Datamart structure id recorded for the template node.
+     * @param integer $controlId Control ID of the inventory data (collection, sample, aliquot) that will be created from the node.
+     */
+    public function defaultValue($nodeId, $datamartStructureId, $controlId)
+    {
+        AppController::addWarningMsg(__('click on submit button of the main form to record the default values'));
+        AppController::forceMsgDisplayInPopup();
+        $structure = $this->TemplateNode->getStructuresForNodeDefaultValuesEntry($datamartStructureId, $controlId);
+        $this->set("structure", $structure);
+        $this->set("nodeId", $nodeId);
+    }
+
+    public function formatedDefaultValue($nodeDatamartStructureId, $nodeControlId)
+    {
+        $defaultValue = $this->request->data;
+        $formatedDefaultValue = $this->TemplateNode->formatTemplateNodeDefaultValuesForDisplayByControlAndDatamartStructureIdId($nodeDatamartStructureId, $nodeControlId, $defaultValue);
+        $this->set("formatedDefaultValue", $formatedDefaultValue);
     }
 }
